+++ src/Serialize.cc
@@ -99,7 +99,6 @@ const std::string SerializeKey::Fixed = "fixed";
 const std::string SerializeKey::True = "true";
 const std::string SerializeKey::Generic = "generic";
 const std::string SerializeKey::Enum = "enum";
-const std::string SerializeKey::Select = "select";
 const std::string SerializeKey::Ref = "ref";
 const std::string SerializeKey::Href = "href";
 const std::string SerializeKey::Path = "path";